// src/modules/content/models/SongDetail.ts
export class SongDetail {
  id?: string;
  praiseChartsId?: string;
  musicBrainzId?: string;
  title?: string;
  artist?: string;
  album?: string;
  language?: string;
  thumbnail?: string;
  releaseDate?: Date;
  bpm?: number;
  keySignature?: string;
  seconds?: number;
  meter?: string;
  tones?: string;
}